pyproject.toml
[project]
name = "patobot"
version = "0.1.0"
description = ""
authors = [
    {name = "PatoProgramador",email = "skpates18@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "discord-py (>=2.5.2,<3.0.0)",
    "py-cord (>=2.6.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "yt-dlp (>=2025.7.21,<2026.0.0)",
    "pynacl (>=1.5.0,<2.0.0)",
    "opuslib (>=3.0.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
invoke = "^2.2.0"

